requirements.txt
flask>=2.2
pandas>=1.3
tensorflow==2.10.0
keras<3.0.0
numpy<2.0
gunicorn>=20.1
pillow>=9.2
